requirements.txt
Flask==3.1.1
numpy==1.26.4
pillow==11.3.0
requests==2.31.0
mediapipe>=0.12.0
supabase==1.1.1
storage3>=0.5.0

# Optional but already included
absl-py==2.3.0
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
flatbuffers==25.2.10
fonttools==4.58.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3